requirements.txt
flask>=3.0
anthropic>=0.100
python-dotenv>=1.0
